+++ lib/ui/settings/SettingsMealPlanPanel.dart
@@ -16,6 +16,7 @@
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:illinois/service/FlexUI.dart';
 import 'package:illinois/ui/settings/SettingsIlliniCashPanel.dart';
 import 'package:rokwire_plugin/service/app_datetime.dart';
 import 'package:illinois/model/IlliniCash.dart';
@@ -696,7 +697,7 @@ class _SettingsMealPlanPanelState extends State<SettingsMealPlanPanel> implement
   }
 
   bool get _canSignIn{
-    return Auth2().privacyMatch(4);
+    return FlexUI().isAuthenticationAvailable;
   }
 
   // NotificationsListener