+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Optimized Groups /user/login API call [#2316](https://github.com/rokwire/illinois-app/issues/2316).
 - Improved scheduling of app review prompt [#2321](https://github.com/rokwire/illinois-app/issues/2321).
+- Removed direct privacy level checks from app code, use FlexUI features instead [#2325](https://github.com/rokwire/illinois-app/issues/2325).
 
 ## [4.1.22] - 2022-08-25
 - Removed any references to PassKit.framework [#1851](https://github.com/rokwire/illinois-app/issues/1851).